+++ src/alley/wp/alleyvate/load.php
@@ -114,6 +114,10 @@ function load(): void {
 			new Features\Disable_Block_Editor_Rest_Api_Preload_Paths(),
 		),
 		new Feature(
+			'noindex_password_protected_posts',
+			new Features\Noindex_Password_Protected_Posts(),
+		),
+		new Feature(
 			'twitter_embeds',
 			new Features\Twitter_Embeds(),
 		),