+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This library adheres to [Semantic Versioning](https://semver.org/) and [Keep a C
 
 ## Unreleased
 
+### Added
+
+* `noindex_password_protected_posts`: Added a feature to add noindex to the robots meta tag content for password-protected posts.
+
 ## 3.6.0
 
 ### Added